Exec approved limit to required quorum and proxies
The exec has approved (on 24 Jan 2019):
A cap on the required quorum of 10 per cause. This is intended to be a temporary measure to help causes function until a better solution is worked out.
Due to the above, to prevent a single member controlling the outcomes of votes, cause members have been limited to holding only two proxy votes.